+++ App/Server/src/bind/bind.c
@@ -6,13 +6,12 @@
 */
 
 #include <stdlib.h>
-#include <stdbool.h>
 
 #include "Server/arguments.h"
 
 bool bind_port(arguments_t *args, char *arg)
 {
-    args->port = atoi(arg);
+    args->port = (u_int16_t) atoi(arg);
     return true;
 }
